+++ test/td_dd_web/controllers/structure_note_controller_test.exs
@@ -2,6 +2,7 @@ defmodule TdDdWeb.StructureNoteControllerTest do
   use TdDdWeb.ConnCase
   use PhoenixSwagger.SchemaTest, "priv/static/swagger.json"
 
+  alias TdCluster.TestHelpers.TdAiMock
   alias TdDd.DataStructures.StructureNote
 
   import TdDd.TestOperators
@@ -82,7 +83,7 @@ defmodule TdDdWeb.StructureNoteControllerTest do
         |> get(Routes.data_structure_note_path(conn, :index, data_structure_id))
         |> json_response(:ok)
 
-      assert ["draft"] == Map.keys(actions)
+      assert ["draft"] ||| Map.keys(actions)
       assert ["deprecated"] == Map.keys(sn_actions)
     end
 
@@ -105,7 +106,9 @@ defmodule TdDdWeb.StructureNoteControllerTest do
 
       assert [] == Map.keys(actions)
       assert [] == Map.keys(v1_actions)
-      assert ["deleted", "edited", "pending_approval", "published"] ||| Map.keys(v2_actions)
+
+      assert ["deleted", "edited", "pending_approval", "published", "ai_suggestions"] |||
+               Map.keys(v2_actions)
     end
 
     @tag authentication: [role: "admin"]
@@ -152,7 +155,143 @@ defmodule TdDdWeb.StructureNoteControllerTest do
         |> get(Routes.data_structure_note_path(conn, :index, data_structure_id))
         |> json_response(:ok)
 
-      assert ["draft"] == Map.keys(actions)
+      assert ["draft"] ||| Map.keys(actions)
+    end
+
+    @tag authentication: [role: "admin"]
+    test "ai_suggestions action available if there are ai_suggestion fields in template and TdAi.ResourceMapping available",
+         %{
+           conn: conn
+         } do
+      template = %{
+        id: System.unique_integer([:positive]),
+        label: "suggestions_test",
+        name: "suggestions_test",
+        scope: "dd",
+        content: [
+          %{
+            "name" => "Identifier Template",
+            "fields" => [
+              %{
+                "cardinality" => "1",
+                "description" => "field description",
+                "label" => "suggestion_field",
+                "name" => "suggestion_field",
+                "type" => "string",
+                "ai_suggestion" => true
+              },
+              %{
+                "cardinality" => "1",
+                "label" => "not_suggestion_field",
+                "name" => "not_suggestion_field",
+                "type" => "string"
+              }
+            ]
+          }
+        ]
+      }
+
+      TdAiMock.available_resource_mapping(&Mox.expect/4, [], {:ok, true})
+
+      %{id: template_id} = CacheHelpers.insert_template(template)
+      CacheHelpers.insert_structure_type(name: template.name, template_id: template_id)
+
+      %{id: data_structure_id} = data_structure = insert(:data_structure)
+
+      insert(:data_structure_version,
+        data_structure: data_structure,
+        type: template.name
+      )
+
+      %{
+        "data" => [],
+        "_actions" => actions
+      } =
+        conn
+        |> get(Routes.data_structure_note_path(conn, :index, data_structure_id))
+        |> json_response(:ok)
+
+      assert actions |> Map.keys() |> Enum.member?("ai_suggestions")
+    end
+
+    @tag authentication: [role: "admin"]
+    test "ai_suggestions action not available if no ai_suggestion fields in template", %{
+      conn: conn
+    } do
+      %{id: template_id} = template = CacheHelpers.insert_template()
+      CacheHelpers.insert_structure_type(name: template.name, template_id: template_id)
+
+      %{id: data_structure_id} = data_structure = insert(:data_structure)
+
+      insert(:data_structure_version,
+        data_structure: data_structure,
+        type: template.name
+      )
+
+      %{
+        "data" => [],
+        "_actions" => actions
+      } =
+        conn
+        |> get(Routes.data_structure_note_path(conn, :index, data_structure_id))
+        |> json_response(:ok)
+
+      refute actions |> Map.keys() |> Enum.member?("ai_suggestions")
+    end
+
+    @tag authentication: [role: "admin"]
+    test "ai_suggestions action not available if TdAi.ResourceMapping not available", %{
+      conn: conn
+    } do
+      template = %{
+        id: System.unique_integer([:positive]),
+        label: "suggestions_test",
+        name: "suggestions_test",
+        scope: "dd",
+        content: [
+          %{
+            "name" => "Identifier Template",
+            "fields" => [
+              %{
+                "cardinality" => "1",
+                "description" => "field description",
+                "label" => "suggestion_field",
+                "name" => "suggestion_field",
+                "type" => "string",
+                "ai_suggestion" => true
+              },
+              %{
+                "cardinality" => "1",
+                "label" => "not_suggestion_field",
+                "name" => "not_suggestion_field",
+                "type" => "string"
+              }
+            ]
+          }
+        ]
+      }
+
+      %{id: template_id} = CacheHelpers.insert_template(template)
+      CacheHelpers.insert_structure_type(name: template.name, template_id: template_id)
+
+      %{id: data_structure_id} = data_structure = insert(:data_structure)
+
+      insert(:data_structure_version,
+        data_structure: data_structure,
+        type: template.name
+      )
+
+      TdAiMock.available_resource_mapping(&Mox.expect/4, [], {:ok, false})
+
+      %{
+        "data" => [],
+        "_actions" => actions
+      } =
+        conn
+        |> get(Routes.data_structure_note_path(conn, :index, data_structure_id))
+        |> json_response(:ok)
+
+      refute actions |> Map.keys() |> Enum.member?("ai_suggestions")
     end
 
     @tag authentication: [role: "admin"]
@@ -726,6 +865,64 @@ defmodule TdDdWeb.StructureNoteControllerTest do
                Routes.data_structure_note_path(conn, :delete, data_structure_id, structure_note)
              )
              |> response(:forbidden)
+    end
+  end
+
+  describe "note_suggestions" do
+    @tag authentication: [role: "admin"]
+    test "retrieves suggestions for the requested fields", %{conn: conn} do
+      template = %{
+        id: System.unique_integer([:positive]),
+        label: "suggestions_test",
+        name: "suggestions_test",
+        scope: "dd",
+        content: [
+          %{
+            "name" => "Identifier Template",
+            "fields" => [
+              %{
+                "cardinality" => "1",
+                "label" => "suggestion_field",
+                "name" => "suggestion_field",
+                "type" => "string",
+                "ai_suggestion" => true
+              },
+              %{
+                "cardinality" => "1",
+                "label" => "not_suggestion_field",
+                "name" => "not_suggestion_field",
+                "type" => "string"
+              }
+            ]
+          }
+        ]
+      }
+
+      %{id: template_id, name: template_name} = CacheHelpers.insert_template(template)
+      CacheHelpers.insert_structure_type(name: template_name, template_id: template_id)
+
+      %{id: data_structure_id} = data_structure = insert(:data_structure)
+
+      insert(:data_structure_version,
+        data_structure: data_structure,
+        type: template_name
+      )
+
+      args = {"data_structure", data_structure_id, %{name: "field"}}
+
+      TdAiMock.resource_field_completion(&Mox.expect/4, args, {:ok, %{}})
+
+      assert %{} ==
+               conn
+               |> get(
+                 Routes.data_structure_structure_note_path(
+                   conn,
+                   :note_suggestions,
+                   data_structure_id
+                 )
+               )
+               |> json_response(:ok)
+               |> Map.get("data")
     end
   end