+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [TD-5390] Validate inplementation key unique when update implementation
 - [TD-4304] Added structures links aggregation in implementation search
+- [TD-6140] Support for AI suggestions on StructureNotes
 
 ### Changed
 
@@ -40,7 +41,14 @@
 - [TD-6088] Improve implementations reindex
 
 ## [5.16.2] 2023-10-19
+
+<<<<<<< HEAD
+
 ## [5.16.1] 2023-10-18
+
+=======
+
+> > > > > > > 81e6d56d160f55ad43e453dfbe0122100524649d
 
 ### Fixed